--- ModelCards/KNNClassifier_v1.md
@@ -1,22 +1,0 @@
-# KNN Classifier - v1
-## General Information 
-- Developed by: donat
-- Model Type: KNN Classifier
-- sklearn: 1.5.2
-- Python Version: 3.12.7
-## Training Details
-- Dataset: brest_cancer.csv
-- Parameters:
-   - `algorithm` auto
-   - `metric` euclidean
-   - `n_neighbors` 5
-   - `weights` uniform
-
-- Training started at: 11:04:11 2024-10-09
-- Training ended at: 11:04:20 2024-10-09
-## Evaluation
-   - `Accuracy` 0.9649122807017544
-   - `F1_macro score` 0.9626596790042581
-   - `F1_micro score` 0.9649122807017544
-   - `Precision` 0.9649122807017544
-   - `Recall` 0.9649122807017544
