--- ModelCards/RandomForestClassifier_v3.md
@@ -1,23 +1,0 @@
-# Random Forest Classifier - v3
-## General Information 
-- Developed by: donat
-- Model Type: Random Forest Classifier
-- sklearn: 1.5.2
-- Python Version: 3.12.7
-## Training Details
-- Dataset: brest_cancer.csv
-- Parameters:
-   - `criterion` gini
-   - `max_depth` None
-   - `min_samples_leaf` 1
-   - `min_samples_split` 5
-   - `n_estimators` 50
-
-- Training started at: 11:12:56 2024-10-09
-- Training ended at: 11:16:35 2024-10-09
-## Evaluation
-   - `Accuracy` 0.9649122807017544
-   - `F1_macro score` 0.9623015873015872
-   - `F1_micro score` 0.9649122807017544
-   - `Precision` 0.9649122807017544
-   - `Recall` 0.9649122807017544
