--- ModelCards/RandomforestwithkMeans_v15.md
@@ -1,23 +1,0 @@
-# Random forest with kMeans - v15
-## General Information 
-- Developed by: donat
-- Model Type: Random forest with kMeans
-- sklearn: 1.5.2
-- Python Version: 3.12.7
-## Training Details
-- Dataset: dataset
-- Parameters:
-   - `criterion` gini
-   - `max_depth` None
-   - `min_samples_leaf` 1
-   - `min_samples_split` 2
-   - `n_estimators` 50
-
-- Training started at: 17:17:38 2024-10-03
-- Training ended at: 17:22:50 2024-10-03
-## Evaluation
-   - `Accuracy` 0.9649122807017544
-   - `F1_macro score` 0.9623015873015872
-   - `F1_micro score` 0.9649122807017544
-   - `Precision` 0.9649122807017544
-   - `Recall` 0.9649122807017544
