+++ ModelTracker/Utils/utility.py
@@ -1,7 +1,6 @@
-from datetime import datetime
 from Utils.kmeans import kMeans
 from Dataset.dataset import Dataset
-import mlflow, json, numpy as np, pandas as pd
+import mlflow, numpy as np, pandas as pd
 
 def preprocessing(dataset : Dataset, cluster = False):
     dataset.dropDatasetColumns(["id"])
@@ -19,23 +18,6 @@ def preprocessing(dataset : Dataset, cluster = False):
     dataset.dropDatasetColumns(['Unnamed: 32'])
     return dataset
 
-def convertTime(unixTime):
-    return datetime.fromtimestamp(unixTime/1000.0).strftime('%H:%M:%S %Y-%m-%d')
-
-def extractInfoTags(tags):   
-    data_tags = json.loads(tags.get('mlflow.log-model.history', ''))
-    flavors = data_tags[0]['flavors']
-    py_version = flavors['python_function']['python_version']
-    lib = str([key for key in flavors.keys() if key != 'python_function'][0])
-    lib_version = flavors[lib].get(f'{lib}_version')
-    return py_version, lib, lib_version
-
-def extratDatasetName(data):
-    dataString = str(data)
-    start = dataString.find("name='") + len("name='")
-    end = dataString.find("'", start)
-    return dataString[start:end]
-
 def inferModel(dataset : Dataset, modelInfo, X_test, y_test):
     loadedModel = mlflow.pyfunc.load_model(modelInfo.model_uri)
     predictions = loadedModel.predict(X_test)
@@ -51,9 +33,3 @@ def inferModel(dataset : Dataset, modelInfo, X_test, y_test):
     result["actual_class"] = y_test
     result["predicted_class"] = predictions
     result.sample(100).to_csv('Evaluation/predictions.csv', index=False)
-    
-def organize(text, info):
-    string = f"{text}\n"
-    for key, value in info.items():
-        string += (f"   - `{key}` {value}\n")
-    return string
